requirements.txt
numpy>=1.26
pandas>=2.2
scipy>=1.12
scikit-learn>=1.4
pyyaml>=6.0
yfinance>=0.2.54
pyarrow>=15.0
requests>=2.31
matplotlib>=3.8
